requirements.txt
discord.py
apify-client